+++ energy_hub/__init__.py
@@ -1,0 +1,1 @@
+from energy_hub.model import EHubModel
