+++ energy_hub/input_data.py
